+++ messages/en.json
@@ -209,9 +209,9 @@
     "projects": "Projects",
     "blog": "Blog",
     "privacy_policy": "Privacy Policy",
-    "site_use_policy": "Website Usage Policy",
+    "site_use_policy": "Terms of Use",
     "regulations": "Regulations",
-    "reporting": "Reporting",
+    "reporting": "Reports",
     "development": "Development",
     "rights": "All rights reserved"
   },
