+++ messages/ua.json
@@ -210,7 +210,7 @@
     "projects": "Проєкти",
     "blog": "Блог",
     "privacy_policy": "Політика конфіденційності",
-    "site_use_policy": "Політика Користування сайтом",
+    "site_use_policy": "Політика користування сайтом",
     "regulations": "Статут",
     "reporting": "Звітність",
     "development": "Розробка",
