+++ src/components/modals/RegistrationFormModal/FormMentor/options.js
@@ -1,0 +1,17 @@
+export const optionsSpec = [
+  { id: "uiuxDesigner", label: "UI/UX Designer" },
+  { id: "frontend", label: "Frontend" },
+  { id: "backend", label: "Backend" },
+  { id: "fullstackEngineer", label: "Fullstack Engineer" },
+  { id: "qaManualEngineer", label: "QA Manual Engineer" },
+  { id: "projectManager", label: "Project Manager" },
+];
+
+// export const optionsTime = [
+//   { id: "t9001200", label: "9.00-12.00" },
+//   { id: "t12001500", label: "12.00-15.00" },
+//   { id: "t18002100", label: "18.00-21.00" },
+//   { id: "anytime", label: "anytime" },
+// ];
+
+//export const optionsPersonalData = [{ id: "personaldata", label: t("permit") }];
