+++ src/components/shared/CloseBtn/CloseBtn.module.scss
@@ -12,8 +12,9 @@
       height: 32px;
     }
   }
-
-  &:hover{
-    opacity: .7;
+  @media (any-hover: hover){
+    &:hover{
+      opacity: .7;
+    }
   }
 }
