+++ src/components/shared/LangDropdown/LangDropdown.module.scss
@@ -50,8 +50,11 @@
     text-align: start;
     transition: all 0.3s;
   }
-
-  .options_item_btn:hover {
-    color: var.$primary-hover;
+  
+  @media (any-hover: hover){
+    .options_item_btn:hover {
+      color: var.$primary-hover;
+    }
   }
+
 }
