+++ src/components/shared/Logo/Logo.module.scss
@@ -8,8 +8,10 @@
     height: 60px;
     fill: var.$secondary-white;
   }
-  &:hover .logo {
-    fill: var.$primary-hover;
+  @media (any-hover: hover){
+    &:hover .logo {
+      fill: var.$primary-hover;
+    }
   }
 }
 
@@ -21,9 +23,10 @@
     height: 120px;
     fill: var.$secondary-white;
   }
-
-  &:hover .logo {
-    fill: var.$primary-hover;
+  @media (any-hover: hover){
+    &:hover .logo {
+      fill: var.$primary-hover;
+    }
   }
 }
